requirements.txt
# Core Dependencies
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0
scikit-learn==1.4.0

# Data Fetching
yfinance==0.2.28
alpha-vantage==2.3.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Technical Analysis
ta==0.10.2
pandas-ta==0.3.14b0

# Machine Learning
xgboost==2.0.3
lightgbm==4.3.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# API & WebSocket
requests==2.31.0
websocket-client==1.6.1
python-socketio==5.8.0
pycryptodome==3.18.0

# Performance & Optimization
numba==0.59.0
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0
loguru==0.7.0

# Data Storage
redis==4.6.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# AlgoLab Integration
pytz==2023.3
tzdata==2023.3